refact: restructure parser (split into tiny submodules)
