+++ src/parser/parser/declaration.rs
@@ -1,0 +1,7 @@
+pub struct Parser;
+
+impl Parser {
+    pub fn new() -> Parser {
+        Parser {}
+    }
+}
