+++ src/parser/parser/mod.rs
@@ -1,0 +1,5 @@
+mod declaration;
+mod multi_module;
+mod single_module;
+
+pub use declaration::Parser;
