+++ src/parser/parser/single_module/mod.rs
@@ -1,0 +1,5 @@
+mod implementation;
+use lalrpop_util::lalrpop_mod;
+lalrpop_mod!(cds, "/parser/parser/single_module/cds.rs");
+#[cfg(test)]
+mod tests;
